requirements.txt
numpy==1.26.4
psycopg[binary]==3.1.18
psycopg[pool]==3.2.1
python-dotenv==1.0.1